Added features for developers
I can't exactly see if Ellie is joining or leaving servers. Nor, do I know the size at which Ellie's audience is. 
To help cope with this, I've added some features where it automatically logs this like server size and id, every time it joins a server.
It will also log if it is kicked/leave a server.